refactor: 앱 셸 구조 정리와 런타임 정책 메타화 (#121)
* fix: 홈 접근성 개선

- viewport 확대 제한 제거

- 검색 버튼과 입력창에 접근성 라벨 추가

- GNB와 캐러셀 컨트롤 접근성 속성 보강

- 터치 타겟 크기와 텍스트 대비 개선

* refactor: 이미지 카드 캐러셀 도트 구조 정리

- 캐러셀 도트를 클릭형 버튼에서 표시용 인디케이터로 변경

- 현재 이미지 인덱스 기준으로 활성 상태만 시각화하도록 정리

- 도트 표시 모드를 static, hidden으로 분리

* feat: 홈 추천 업체 카드 도트 숨김 적용

- 홈 추천 업체 카드에 hidden 도트 모드 적용

- 다른 카드 영역은 기본 static 도트 유지

* refactor: 로그인 모달 전역 로딩 구조 개선

- 로그인 모달을 gate 기반 지연 로드로 전환

- _app에서 정적 import 대신 필요 시 모달 청크를 로드하도록 변경

* refactor: 챗봇 런처와 런타임 분리

- 챗봇 플로팅 런처와 무거운 모달 런타임을 분리

- 사용자 상호작용 시점에 챗봇 런타임 청크를 지연 로드하도록 변경

- hover, focus, touch 시 prefetch를 추가

* refactor: GTM 로딩 시점 최적화

- GTM 로더를 _document 즉시 주입에서 lazyOnload로 전환

- dataLayer 및 consent 기본값 초기화는 유지

- 초기 로드 시 3rd-party 스크립트 비용을 줄이도록 조정

* refactor: 앱 셸 구조와 런타임 정책 정리

- _app를 앱 셸 조립 역할만 하도록 정리

- 메타, 프로바이더, 런타임, 페이지 렌더 분기를 app-shell로 분리

- 페이지 정책 기반으로 analytics와 auth sync 동작 범위를 제어

* refactor: 인증과 사용자 런타임 로딩 구조 개선

- 인증 bootstrap 중복 실행을 방지하도록 하드닝

- 로그인 모달의 LoginForm 런타임을 지연 로드로 분리

- 챗봇 런타임 prefetch를 인증 사용자 기준으로 제한

- 마케팅 동의 상태를 auth store와 GTM 동기화 흐름에 맞게 정리

* refactor: 서버 프록시 유틸 구조 정리

- agent chat 프록시를 단일 경로 매핑 구조로 유지

- JSON 프록시 요청 로직을 공통 유틸 형태로 정리
