+++ src/components/company/company-card/index.tsx
@@ -4,6 +4,7 @@ import { Location, ChevronLeftWhite } from '@/icons';
 import NextImage from 'next/image';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { normalizeSafeImageSrc, shouldBypassNextImageOptimization } from '@/utils/image';
+import { useTranslations } from 'next-intl';
 import {
   wrapper,
   wrapperFixedHeight,
@@ -42,6 +43,7 @@ interface Props {
   size?: 'default' | 'compact';
   images?: string[]; // 여러 이미지 배열 추가
   isExclusive?: boolean;
+  carouselDotsMode?: 'static' | 'hidden';
   onClick: (clinicId: number) => void;
 }
 
@@ -56,8 +58,10 @@ export function CompanyCard({
   size = 'default',
   images = [],
   isExclusive = false,
+  carouselDotsMode = 'static',
   onClick,
 }: Props) {
+  const t = useTranslations('common');
   const [imageError, setImageError] = useState(false);
   const [currentImageIndex, setCurrentImageIndex] = useState(0);
   const ignoreClickRef = useRef(false);
@@ -93,10 +97,6 @@ export function CompanyCard({
     if (currentImageIndex < imageList.length) return;
     setCurrentImageIndex(0);
   }, [currentImageIndex, imageList.length]);
-
-  const handleDotClick = (index: number) => {
-    setCurrentImageIndex(index);
-  };
 
   const goPrevImage = () => {
     setCurrentImageIndex((prev) => (prev > 0 ? prev - 1 : imageList.length - 1));
@@ -183,32 +183,32 @@ export function CompanyCard({
                 <img src="/default.png" alt="기본 이미지" css={carouselImage} />
               )}
               <button
+                type="button"
                 css={[carouselNavButton, carouselNavLeft]}
                 onClick={handlePrevImage}
-                aria-label="이전 이미지"
+                aria-label={t('button.previous')}
               >
                 <ChevronLeftWhite width={32} height={34} />
               </button>
               <button
+                type="button"
                 css={[carouselNavButton, carouselNavRight]}
                 onClick={handleNextImage}
-                aria-label="다음 이미지"
+                aria-label={t('button.next')}
               >
                 <ChevronLeftWhite width={32} height={34} style={{ transform: 'rotate(180deg)' }} />
               </button>
             </div>
-            <div css={carouselDots}>
-              {imageList.map((_, index) => (
-                <button
-                  key={index}
-                  css={index === currentImageIndex ? carouselDotActive : carouselDot}
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    handleDotClick(index);
-                  }}
-                />
-              ))}
-            </div>
+            {carouselDotsMode === 'static' && (
+              <div css={carouselDots} aria-hidden="true">
+                {imageList.map((_, index) => (
+                  <span
+                    key={index}
+                    css={index === currentImageIndex ? carouselDotActive : carouselDot}
+                  />
+                ))}
+              </div>
+            )}
           </div>
         ) : (
           <>
