+++ src/components/search-bar/index.styles.ts
@@ -15,9 +15,9 @@ export const searchBar = css`
   align-items: center;
 
   width: 100%;
-  height: 42px;
+  height: 48px;
   margin: 2px 20px 12px;
-  padding: 0 16px;
+  padding: 0 12px;
   border: 2px solid ${theme.colors.border_default};
   border-radius: 32px;
 
@@ -26,19 +26,30 @@ export const searchBar = css`
 
 export const input = css`
   width: 100%;
+  min-width: 0;
 
+  color: ${theme.colors.text_primary};
   font-size: ${theme.typo.body_M};
   text-align: left;
 
   ::placeholder {
-    color: ${theme.colors.primary30};
+    color: ${theme.colors.text_tertiary};
   }
 `;
 
 export const button = ({ isLeft }: { isLeft: boolean }) => css`
+  display: flex;
   flex-shrink: 0;
+  align-items: center;
+  justify-content: center;
 
-  width: 24px;
-  height: 24px;
-  margin: ${isLeft ? '0 8px 0 0' : '0 0 0 8px'};
+  width: 44px;
+  height: 44px;
+  margin: ${isLeft ? '0 4px 0 -4px' : '0 -4px 0 4px'};
+  border-radius: 999px;
+
+  &:focus-visible {
+    outline: 2px solid ${theme.colors.primary50};
+    outline-offset: 2px;
+  }
 `;
